+++ mix.exs
@@ -6,6 +6,7 @@ defmodule Sptfy.MixProject do
       app: :sptfy,
       version: "0.1.0",
       elixir: "~> 1.11",
+      elixirc_paths: elixirc_paths(Mix.env()),
       start_permanent: Mix.env() == :prod,
       deps: deps()
     ]
@@ -17,6 +18,9 @@ defmodule Sptfy.MixProject do
       mod: {Sptfy.Application, []}
     ]
   end
+
+  defp elixirc_paths(:test), do: ["lib", "test/support"]
+  defp elixirc_paths(_), do: ["lib"]
 
   defp deps do
     [
