+++ test/sptfy/album_test.exs
@@ -10,7 +10,7 @@ defmodule Sptfy.AlbumTest do
     test "returns list of FullAlbum structs" do
       json = %{"albums" => [Fixtures.full_album()]}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/albums", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/albums", _ -> MockHelpers.response(json) end do
         assert {:ok, [%FullAlbum{}]} = Album.get_albums("token", ids: [1, 2, 3])
       end
     end
@@ -18,7 +18,7 @@ defmodule Sptfy.AlbumTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/albums", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/albums", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Album.get_albums("token", ids: [1, 2, 3])
       end
     end
@@ -26,7 +26,7 @@ defmodule Sptfy.AlbumTest do
 
   describe "get_album/2" do
     test "returns a FullAlbum struct" do
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/albums/abc", _ -> TestHelpers.response(Fixtures.full_album()) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/albums/abc", _ -> MockHelpers.response(Fixtures.full_album()) end do
         assert {:ok, %FullAlbum{}} = Album.get_album("token", id: "abc")
       end
     end
@@ -34,7 +34,7 @@ defmodule Sptfy.AlbumTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/albums/abc", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/albums/abc", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Album.get_album("token", id: "abc")
       end
     end
@@ -44,7 +44,7 @@ defmodule Sptfy.AlbumTest do
     test "returns a Paging struct" do
       json = Fixtures.paging(Fixtures.simplified_track())
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/albums/abc/tracks", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/albums/abc/tracks", _ -> MockHelpers.response(json) end do
         assert {:ok, %Paging{items: [%SimplifiedTrack{}]}} = Album.get_album_tracks("token", id: "abc")
       end
     end
@@ -52,7 +52,7 @@ defmodule Sptfy.AlbumTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/albums/abc/tracks", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/albums/abc/tracks", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Album.get_album_tracks("token", id: "abc")
       end
     end
