+++ test/sptfy/episode_test.exs
@@ -10,7 +10,7 @@ defmodule Sptfy.EpisodeTest do
     test "returns list of FullEpisode structs" do
       json = %{"episodes" => [Fixtures.full_episode()]}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/episodes", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/episodes", _ -> MockHelpers.response(json) end do
         assert {:ok, [%FullEpisode{}]} = Episode.get_episodes("token", ids: [1, 2, 3])
       end
     end
@@ -18,7 +18,7 @@ defmodule Sptfy.EpisodeTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/episodes", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/episodes", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Episode.get_episodes("token", ids: [1, 2, 3])
       end
     end
@@ -26,7 +26,7 @@ defmodule Sptfy.EpisodeTest do
 
   describe "get_episode/2" do
     test "returns a FullEpisode struct" do
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/episodes/abc", _ -> TestHelpers.response(Fixtures.full_episode()) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/episodes/abc", _ -> MockHelpers.response(Fixtures.full_episode()) end do
         assert {:ok, %FullEpisode{}} = Episode.get_episode("token", id: "abc")
       end
     end
@@ -34,7 +34,7 @@ defmodule Sptfy.EpisodeTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/episodes/abc", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/episodes/abc", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Episode.get_episode("token", id: "abc")
       end
     end
