+++ test/sptfy/follow_test.exs
@@ -8,7 +8,7 @@ defmodule Sptfy.FollowTest do
 
   describe "follow_playlist/2" do
     test "returns :ok" do
-      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc/followers", _, _ -> TestHelpers.response(%{}) end do
+      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc/followers", _, _ -> MockHelpers.response(%{}) end do
         assert :ok == Follow.follow_playlist("token", id: "abc")
       end
     end
@@ -16,7 +16,7 @@ defmodule Sptfy.FollowTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc/followers", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc/followers", _, _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Follow.follow_playlist("token", id: "abc")
       end
     end
@@ -24,7 +24,7 @@ defmodule Sptfy.FollowTest do
 
   describe "unfollow_playlist/2" do
     test "returns :ok" do
-      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/playlists/abc/followers", _, _ -> TestHelpers.response(%{}) end do
+      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/playlists/abc/followers", _, _ -> MockHelpers.response(%{}) end do
         assert :ok == Follow.unfollow_playlist("token", id: "abc")
       end
     end
@@ -32,7 +32,7 @@ defmodule Sptfy.FollowTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/playlists/abc/followers", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/playlists/abc/followers", _, _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Follow.unfollow_playlist("token", id: "abc")
       end
     end
@@ -40,7 +40,7 @@ defmodule Sptfy.FollowTest do
 
   describe "check_playlist_following_state/2" do
     test "returns list of booleans" do
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc/followers/contains", _ -> TestHelpers.response([true]) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc/followers/contains", _ -> MockHelpers.response([true]) end do
         assert {:ok, [true]} == Follow.check_playlist_following_state("token", id: "abc", ids: ["def"])
       end
     end
@@ -48,7 +48,7 @@ defmodule Sptfy.FollowTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc/followers/contains", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc/followers/contains", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Follow.check_playlist_following_state("token", id: "abc", ids: ["def"])
       end
     end
@@ -58,7 +58,7 @@ defmodule Sptfy.FollowTest do
     test "returns a Paging struct" do
       json = %{"artists" => Fixtures.paging(Fixtures.full_artist())}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/following", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/following", _ -> MockHelpers.response(json) end do
         assert {:ok, %Paging{items: [%FullArtist{}]}} = Follow.get_my_following_artists("token")
       end
     end
@@ -66,7 +66,7 @@ defmodule Sptfy.FollowTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/following", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/following", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Follow.get_my_following_artists("token")
       end
     end
@@ -74,7 +74,7 @@ defmodule Sptfy.FollowTest do
 
   describe "follow_users/2" do
     test "returns :ok" do
-      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/me/following", _, _ -> TestHelpers.response(%{}) end do
+      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/me/following", _, _ -> MockHelpers.response(%{}) end do
         assert :ok == Follow.follow_users("token", ids: ["abc"])
       end
     end
@@ -82,7 +82,7 @@ defmodule Sptfy.FollowTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/me/following", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/me/following", _, _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Follow.follow_users("token", ids: ["abc"])
       end
     end
@@ -90,7 +90,7 @@ defmodule Sptfy.FollowTest do
 
   describe "follow_artists/2" do
     test "returns :ok" do
-      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/me/following", _, _ -> TestHelpers.response(%{}) end do
+      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/me/following", _, _ -> MockHelpers.response(%{}) end do
         assert :ok == Follow.follow_artists("token", ids: ["abc"])
       end
     end
@@ -98,7 +98,7 @@ defmodule Sptfy.FollowTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/me/following", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/me/following", _, _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Follow.follow_artists("token", ids: ["abc"])
       end
     end
@@ -106,7 +106,7 @@ defmodule Sptfy.FollowTest do
 
   describe "unfollow_users/2" do
     test "returns :ok" do
-      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/me/following", _, _ -> TestHelpers.response(%{}) end do
+      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/me/following", _, _ -> MockHelpers.response(%{}) end do
         assert :ok == Follow.unfollow_users("token", ids: ["abc"])
       end
     end
@@ -114,7 +114,7 @@ defmodule Sptfy.FollowTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/me/following", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/me/following", _, _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Follow.unfollow_users("token", ids: ["abc"])
       end
     end
@@ -122,7 +122,7 @@ defmodule Sptfy.FollowTest do
 
   describe "unfollow_artists/2" do
     test "returns :ok" do
-      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/me/following", _, _ -> TestHelpers.response(%{}) end do
+      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/me/following", _, _ -> MockHelpers.response(%{}) end do
         assert :ok == Follow.unfollow_artists("token", ids: ["abc"])
       end
     end
@@ -130,7 +130,7 @@ defmodule Sptfy.FollowTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/me/following", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/me/following", _, _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Follow.unfollow_artists("token", ids: ["abc"])
       end
     end
@@ -138,7 +138,7 @@ defmodule Sptfy.FollowTest do
 
   describe "check_my_user_following_state/2" do
     test "returns list of booleans" do
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/following/contains", _ -> TestHelpers.response([true]) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/following/contains", _ -> MockHelpers.response([true]) end do
         assert {:ok, [true]} == Follow.check_my_user_following_state("token", ids: ["def"])
       end
     end
@@ -146,7 +146,7 @@ defmodule Sptfy.FollowTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/following/contains", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/following/contains", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Follow.check_my_user_following_state("token", ids: ["def"])
       end
     end
@@ -154,7 +154,7 @@ defmodule Sptfy.FollowTest do
 
   describe "check_my_artist_following_state/2" do
     test "returns list of booleans" do
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/following/contains", _ -> TestHelpers.response([true]) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/following/contains", _ -> MockHelpers.response([true]) end do
         assert {:ok, [true]} == Follow.check_my_artist_following_state("token", ids: ["def"])
       end
     end
@@ -162,7 +162,7 @@ defmodule Sptfy.FollowTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/following/contains", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/following/contains", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Follow.check_my_artist_following_state("token", ids: ["def"])
       end
     end
