+++ test/sptfy/playlist_test.exs
@@ -10,7 +10,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns a Paging struct" do
       json = Fixtures.paging(Fixtures.simplified_playlist())
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/playlists", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/playlists", _ -> MockHelpers.response(json) end do
         assert {:ok, %Paging{items: [%SimplifiedPlaylist{}]}} = Playlist.get_my_playlists("token")
       end
     end
@@ -18,7 +18,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/playlists", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/me/playlists", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Playlist.get_my_playlists("token")
       end
     end
@@ -28,7 +28,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns a Paging struct" do
       json = Fixtures.paging(Fixtures.simplified_playlist())
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/users/abc/playlists", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/users/abc/playlists", _ -> MockHelpers.response(json) end do
         assert {:ok, %Paging{items: [%SimplifiedPlaylist{}]}} = Playlist.get_user_playlists("token", id: "abc")
       end
     end
@@ -36,7 +36,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/users/abc/playlists", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/users/abc/playlists", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Playlist.get_user_playlists("token", id: "abc")
       end
     end
@@ -44,7 +44,7 @@ defmodule Sptfy.PlaylistTest do
 
   describe "create_user_playlist/2" do
     test "returns a Playlist struct" do
-      with_mock Sptfy.Client.HTTP, post: fn _, "/v1/users/abc/playlists", _, _ -> TestHelpers.response(Fixtures.full_playlist()) end do
+      with_mock Sptfy.Client.HTTP, post: fn _, "/v1/users/abc/playlists", _, _ -> MockHelpers.response(Fixtures.full_playlist()) end do
         assert {:ok, %FullPlaylist{}} = Playlist.create_user_playlist("token", id: "abc", name: "Playlist")
       end
     end
@@ -52,7 +52,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, post: fn _, "/v1/users/abc/playlists", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, post: fn _, "/v1/users/abc/playlists", _, _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Playlist.create_user_playlist("token", id: "abc", name: "Playlist")
       end
     end
@@ -60,7 +60,7 @@ defmodule Sptfy.PlaylistTest do
 
   describe "get_playlist/2" do
     test "returns a FullPlaylist struct" do
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc", _ -> TestHelpers.response(Fixtures.full_playlist()) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc", _ -> MockHelpers.response(Fixtures.full_playlist()) end do
         assert {:ok, %FullPlaylist{}} = Playlist.get_playlist("token", id: "abc")
       end
     end
@@ -68,7 +68,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Playlist.get_playlist("token", id: "abc")
       end
     end
@@ -76,7 +76,7 @@ defmodule Sptfy.PlaylistTest do
 
   describe "update_playlist_details/2" do
     test "returns :ok" do
-      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc", _, _ -> TestHelpers.response(%{}) end do
+      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc", _, _ -> MockHelpers.response(%{}) end do
         assert :ok == Playlist.update_playlist_details("token", id: "abc", name: "Playlist")
       end
     end
@@ -84,7 +84,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc", _, _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Playlist.update_playlist_details("token", id: "abc", name: "Playlist")
       end
     end
@@ -94,7 +94,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns a Paging struct" do
       json = Fixtures.paging(Fixtures.playlist_track())
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc/tracks", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc/tracks", _ -> MockHelpers.response(json) end do
         assert {:ok, %Paging{items: [%PlaylistTrack{}]}} = Playlist.get_playlist_tracks("token", id: "abc")
       end
     end
@@ -102,7 +102,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc/tracks", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc/tracks", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Playlist.get_playlist_tracks("token", id: "abc")
       end
     end
@@ -112,7 +112,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns :ok" do
       json = %{"snapshot_id" => "SNAPSHOT_ID"}
 
-      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc/tracks", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc/tracks", _, _ -> MockHelpers.response(json) end do
         assert :ok == Playlist.replace_tracks("token", id: "abc", uris: ["uri"])
       end
     end
@@ -120,7 +120,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc/tracks", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc/tracks", _, _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Playlist.replace_tracks("token", id: "abc", uris: ["uri"])
       end
     end
@@ -130,7 +130,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns :ok" do
       json = %{"snapshot_id" => "SNAPSHOT_ID"}
 
-      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc/tracks", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc/tracks", _, _ -> MockHelpers.response(json) end do
         assert :ok == Playlist.reorder_tracks("token", id: "abc", range_start: 0, insert_before: 5)
       end
     end
@@ -138,7 +138,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc/tracks", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, put: fn _, "/v1/playlists/abc/tracks", _, _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Playlist.reorder_tracks("token", id: "abc", range_start: 0, insert_before: 5)
       end
     end
@@ -148,7 +148,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns :ok" do
       json = %{"snapshot_id" => "SNAPSHOT_ID"}
 
-      with_mock Sptfy.Client.HTTP, post: fn _, "/v1/playlists/abc/tracks", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, post: fn _, "/v1/playlists/abc/tracks", _, _ -> MockHelpers.response(json) end do
         assert :ok == Playlist.add_tracks("token", id: "abc", uris: ["uri"], position: 1)
       end
     end
@@ -156,7 +156,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, post: fn _, "/v1/playlists/abc/tracks", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, post: fn _, "/v1/playlists/abc/tracks", _, _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Playlist.add_tracks("token", id: "abc", uris: ["uri"], position: 1)
       end
     end
@@ -166,7 +166,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns a list of Image structs" do
       json = [Fixtures.image()]
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc/images", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc/images", _ -> MockHelpers.response(json) end do
         assert {:ok, [%Image{}]} = Playlist.get_cover_images("token", id: "abc")
       end
     end
@@ -174,7 +174,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc/images", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/playlists/abc/images", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Playlist.get_cover_images("token", id: "abc")
       end
     end
@@ -182,7 +182,7 @@ defmodule Sptfy.PlaylistTest do
 
   describe "upload_cover_image/3" do
     test "returns :ok" do
-      with_mock Sptfy.Client.HTTP, put_jpeg: fn _, "/v1/playlists/abc/images", _, _ -> TestHelpers.response(%{}) end do
+      with_mock Sptfy.Client.HTTP, put_jpeg: fn _, "/v1/playlists/abc/images", _, _ -> MockHelpers.response(%{}) end do
         assert :ok == Playlist.upload_cover_image("token", "base64", id: "abc")
       end
     end
@@ -190,7 +190,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, put_jpeg: fn _, "/v1/playlists/abc/images", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, put_jpeg: fn _, "/v1/playlists/abc/images", _, _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Playlist.upload_cover_image("token", "base64", id: "abc")
       end
     end
@@ -200,7 +200,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns :ok" do
       json = %{"snapshot_id" => "SNAPSHOT_ID"}
 
-      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/playlists/abc/tracks", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/playlists/abc/tracks", _, _ -> MockHelpers.response(json) end do
         assert :ok == Playlist.remove_tracks("token", id: "abc", tracks: [%{uri: "uri"}])
       end
     end
@@ -208,7 +208,7 @@ defmodule Sptfy.PlaylistTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/playlists/abc/tracks", _, _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, delete: fn _, "/v1/playlists/abc/tracks", _, _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Playlist.remove_tracks("token", id: "abc", tracks: [%{uri: "uri"}])
       end
     end
