+++ test/sptfy/search_test.exs
@@ -8,7 +8,7 @@ defmodule Sptfy.SearchTest do
 
   describe "search_album/2" do
     test "returns a Paging struct" do
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> TestHelpers.response(paging_json("albums")) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> MockHelpers.response(paging_json("albums")) end do
         assert {:ok, %Paging{items: [%SimplifiedAlbum{}]}} = Search.search_album("token", q: "q")
       end
     end
@@ -16,7 +16,7 @@ defmodule Sptfy.SearchTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Search.search_album("token", q: "q")
       end
     end
@@ -24,7 +24,7 @@ defmodule Sptfy.SearchTest do
 
   describe "search_artist/2" do
     test "returns a Paging struct" do
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> TestHelpers.response(paging_json("artists")) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> MockHelpers.response(paging_json("artists")) end do
         assert {:ok, %Paging{items: [%FullArtist{}]}} = Search.search_artist("token", q: "q")
       end
     end
@@ -32,7 +32,7 @@ defmodule Sptfy.SearchTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Search.search_artist("token", q: "q")
       end
     end
@@ -40,7 +40,7 @@ defmodule Sptfy.SearchTest do
 
   describe "search_playlist/2" do
     test "returns a Paging struct" do
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> TestHelpers.response(paging_json("playlists")) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> MockHelpers.response(paging_json("playlists")) end do
         assert {:ok, %Paging{items: [%SimplifiedPlaylist{}]}} = Search.search_playlist("token", q: "q")
       end
     end
@@ -48,7 +48,7 @@ defmodule Sptfy.SearchTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Search.search_playlist("token", q: "q")
       end
     end
@@ -56,7 +56,7 @@ defmodule Sptfy.SearchTest do
 
   describe "search_episode/2" do
     test "returns a Paging struct" do
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> TestHelpers.response(paging_json("episodes")) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> MockHelpers.response(paging_json("episodes")) end do
         assert {:ok, %Paging{items: [%SimplifiedEpisode{}]}} = Search.search_episode("token", q: "q")
       end
     end
@@ -64,7 +64,7 @@ defmodule Sptfy.SearchTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Search.search_episode("token", q: "q")
       end
     end
@@ -72,7 +72,7 @@ defmodule Sptfy.SearchTest do
 
   describe "search_show/2" do
     test "returns a Paging struct" do
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> TestHelpers.response(paging_json("shows")) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> MockHelpers.response(paging_json("shows")) end do
         assert {:ok, %Paging{items: [%SimplifiedShow{}]}} = Search.search_show("token", q: "q")
       end
     end
@@ -80,7 +80,7 @@ defmodule Sptfy.SearchTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Search.search_show("token", q: "q")
       end
     end
@@ -88,7 +88,7 @@ defmodule Sptfy.SearchTest do
 
   describe "search_track/2" do
     test "returns a Paging struct" do
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> TestHelpers.response(paging_json("tracks")) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> MockHelpers.response(paging_json("tracks")) end do
         assert {:ok, %Paging{items: [%FullTrack{}]}} = Search.search_track("token", q: "q")
       end
     end
@@ -96,7 +96,7 @@ defmodule Sptfy.SearchTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/search", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Search.search_track("token", q: "q")
       end
     end
