+++ test/sptfy/show_test.exs
@@ -10,7 +10,7 @@ defmodule Sptfy.ShowTest do
     test "returns list of SimplifiedShow structs" do
       json = %{"shows" => [Fixtures.simplified_show()]}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/shows", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/shows", _ -> MockHelpers.response(json) end do
         assert {:ok, [%SimplifiedShow{}]} = Show.get_shows("token", ids: [1, 2, 3])
       end
     end
@@ -18,7 +18,7 @@ defmodule Sptfy.ShowTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/shows", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/shows", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Show.get_shows("token", ids: [1, 2, 3])
       end
     end
@@ -26,7 +26,7 @@ defmodule Sptfy.ShowTest do
 
   describe "get_show/2" do
     test "returns a FullShow struct" do
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/shows/abc", _ -> TestHelpers.response(Fixtures.full_show()) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/shows/abc", _ -> MockHelpers.response(Fixtures.full_show()) end do
         assert {:ok, %FullShow{}} = Show.get_show("token", id: "abc")
       end
     end
@@ -34,7 +34,7 @@ defmodule Sptfy.ShowTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/shows/abc", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/shows/abc", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Show.get_show("token", id: "abc")
       end
     end
@@ -44,7 +44,7 @@ defmodule Sptfy.ShowTest do
     test "returns a Paging struct" do
       json = Fixtures.paging(Fixtures.simplified_episode())
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/shows/abc/episodes", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/shows/abc/episodes", _ -> MockHelpers.response(json) end do
         assert {:ok, %Paging{items: [%SimplifiedEpisode{}]}} = Show.get_episodes("token", id: "abc")
       end
     end
@@ -52,7 +52,7 @@ defmodule Sptfy.ShowTest do
     test "returns Error struct on error" do
       json = %{"error" => %{"message" => "Oops", "status" => 401}}
 
-      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/shows/abc/episodes", _ -> TestHelpers.response(json) end do
+      with_mock Sptfy.Client.HTTP, get: fn _, "/v1/shows/abc/episodes", _ -> MockHelpers.response(json) end do
         assert {:error, %Sptfy.Object.Error{message: "Oops", status: 401}} = Show.get_episodes("token", id: "abc")
       end
     end
