+++ test/support/mock_helpers.ex
@@ -1,0 +1,7 @@
+defmodule MockHelpers do
+  def response(json) do
+    body = json |> Jason.encode!()
+
+    {:ok, %Finch.Response{body: body}}
+  end
+end
